requirements.txt
Django==1.8.15
Django-Select2==4.3.2
Pillow==3.3.1
Unidecode==0.04.19
amqp==1.4.9
anyjson==0.3.3
argparse==1.2.1
billiard==3.3.0.23
celery==3.1.24
cffi==1.8.3
cmsplugin-filer==1.1.3
cryptography==1.5.2
dj-database-url==0.4.1
django-appconf==1.0.2
django-celery==3.1.17
django-classy-tags==0.8.0
django-cors-headers==1.2.2
django-filer==1.2.5
django-formtools==1.0
django-grappelli==2.8.2
django-mptt==0.8.6
django-polymorphic==0.8.1
django-push-notifications==1.4.1
django-pusherable==0.1.0
django-sekizai==0.10.0
django-treebeard==4.0.1
djangorestframework==3.4.7
easy-thumbnails==2.3
enum34==1.1.6
google-api-python-client==1.5.3
html5lib==0.9999999
httplib2==0.9.2
idna==2.1
ipaddress==1.0.17
kombu==3.0.37
ndg-httpsclient==0.4.2
oauth2client==3.0.0
pusher==1.5
pyOpenSSL==16.2.0
pyasn1==0.1.9
pyasn1-modules==0.0.8
pycparser==2.16
pytz==2016.7
redis==2.10.5
requests==2.11.1
rsa==3.4.2
schedule==0.3.2
simplejson==3.8.2
six==1.10.0
tzlocal==1.2.2
uritemplate==0.6
urllib3==1.18
wsgiref==0.1.2
